GPT. Нейросети спешат на помощь. [#505156]

+++ src/test/java/ru/job4j/calculator/FitTest.java
@@ -8,7 +8,7 @@ class FitTest {
 
     @Test
     void whenMan180Then92() {
-        short input = 180;
+        int input = 180;
         double expected = 92;
         double output = Fit.manWeight(input);
         assertThat(output).isEqualTo(expected, withPrecision(0.01));
@@ -16,7 +16,7 @@ class FitTest {
 
     @Test
     void whenWoman170Then69() {
-        short input = 170;
+        int input = 170;
         double expected = 69;
         double output = Fit.womanWeight(input);
         assertThat(output).isEqualTo(expected, withPrecision(0.01));
